Add files via upload
In this project, I took the data on housing prices from Kaggle with various features. I did feature engineering to find the important features and applied advanced regression algorithms to it to predict the housing price. I found the cost function of each model and compared them to find a generalized model.