Add files via upload
In this project, I took the credit card transaction data from Kaggle to develop a Machine Learning model to predict the transaction as fraud and non-fraud transactions. As the data set is highly imbalanced, I employed a technique of random under-sampling and used ML algorithms such as Logistic regression and Support Vector Classifier to develop a model.